requirements.txt
matplotlib
rapidfuzz
pandas
discord
pyinstaller
pyyaml